styled the login page and made some changes

+++ uploads/views.py
@@ -4,14 +4,16 @@ from django.http import HttpResponseRedirect,Http404
 from .forms import PostForm
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
 from django.db.models import Q
+from django.contrib.auth.decorators import login_required
 
 
 # Create your views here.
 def upload_view(request):
 	return render(request,'base.html',context)
 
+@login_required
 def post_list(request):
-	queryset_list = upload.objects.filter(user=request.user) # Get all database querysets
+	queryset_list = upload.objects.all() # Get all database querysets
 	queryset = upload.objects.all()
 
 	query = request.GET.get("q")
@@ -42,6 +44,7 @@ def post_detail(request,id=None):
 	context = {"title":instance.title,"instance":instance}
 	return render(request,"post_detail.html",context)
 
+@login_required
 def create_post(request):
 	# if not request.user.is_staff or not request.user.is_superuser:
 	# 	raise Http404
@@ -58,8 +61,11 @@ def create_post(request):
 
 def delete_post(request,id=None):
 	instance = get_object_or_404(upload,id=id)
-	instance.delete()
-	return redirect('/uploads/list/')
+	if instance.user == request.user:
+		instance.delete()
+		return redirect('/uploads/list/')
+	else:
+		return render(request,"cantdel.html",{})
 
 def update_post(request):
 	form = PostForm(request.POST or None)
@@ -71,9 +77,31 @@ def update_post(request):
 	  "form" : form
 	}
 	return render(request,'form.html',context)
+def my_posts(request):
+	queryset_list = upload.objects.filter(user=request.user) # Get all database querysets
+	queryset = upload.objects.all()
 
+	query = request.GET.get("q")
+	if query:
+		queryset_list = queryset_list.filter(
+			Q(title__icontains=query) |
+            Q(description__icontains=query) |
+            Q(user__first_name__icontains=query) |
+            Q(user__last_name__icontains=query)
+			).distinct()
 
-def read_more(request):
-	return render(request,'read_more.html',{})
+	paginator = Paginator(queryset_list, 2) # Show 2 contacts per page
+
+	page = request.GET.get('page')
+	try:
+	    queryset = paginator.page(page)
+	except PageNotAnInteger:
+	    # If page is not an integer, deliver first page.
+	    queryset = paginator.page(1)
+	except EmptyPage:
+	    # If page is out of range (e.g. 9999), deliver last page of results.
+	    queryset = paginator.page(paginator.num_pages)
+	context = {"object_list" : queryset}
+	return render(request,'post_list.html',context)
